+++ build.sh
@@ -5,5 +5,6 @@ pip install -U -r requirements.txt
 cd docs && make html && cp -r build/html/* ../../checkdmarc-docs/
 cd ..
 rm -rf dist/ build/
-python setup.py bdist_wheel
+python3 setup.py sdist
+python3 setup.py bdist_wheel